Split up the capacity transactions test
