+++ e2e/capacity/transactions.test.ts
@@ -761,113 +761,4 @@ describe('Capacity Transactions', function () {
       });
     });
   });
-
-  describe('pay_with_capacity_batch_all', function () {
-    let capacityProviderKeys: KeyringPair;
-    let capacityProvider: u64;
-    let schemaId: u16;
-    let defaultPayload: AddProviderPayload;
-    const amountStaked = 9n * DOLLARS;
-
-    beforeEach(async function () {
-      capacityProviderKeys = createKeys('CapacityProviderKeys');
-      capacityProvider = await createMsaAndProvider(
-        fundingSource,
-        capacityProviderKeys,
-        'CapacityProvider',
-        FUNDS_AMOUNT
-      );
-      defaultPayload = {
-        authorizedMsaId: capacityProvider,
-        schemaIds: [schemaId],
-      };
-    });
-
-    it('successfully pays with Capacity for a batch of eligible transactions - [createSponsoredAccountWithDelegation, claimHandle]', async function () {
-      await assert.doesNotReject(stakeToProvider(fundingSource, fundingSource, capacityProvider, amountStaked));
-
-      const addProviderPayload = await generateDelegationPayload({ ...defaultPayload });
-      const addProviderData = ExtrinsicHelper.api.registry.createType('PalletMsaAddProvider', addProviderPayload);
-      const delegatorKeys = createKeys('delegatorKeys');
-      const createSponsoredAccountWithDelegation = ExtrinsicHelper.api.tx.msa.createSponsoredAccountWithDelegation(
-        delegatorKeys.publicKey,
-        signPayloadSr25519(delegatorKeys, addProviderData),
-        addProviderPayload
-      );
-
-      const handle = getTestHandle();
-      const handle_vec = new Bytes(ExtrinsicHelper.api.registry, handle);
-      const expiration = (await getBlockNumber()) + 5;
-      const handlePayload = {
-        baseHandle: handle_vec,
-        expiration: expiration,
-      };
-      const claimHandlePayload: any = ExtrinsicHelper.api.registry.createType(
-        'CommonPrimitivesHandlesClaimHandlePayload',
-        handlePayload
-      );
-      const claimHandleProof = {
-        Sr25519: u8aToHex(delegatorKeys.sign(u8aWrapBytes(claimHandlePayload.toU8a()))),
-      };
-
-      const claimHandle = ExtrinsicHelper.api.tx.handles.claimHandle(
-        delegatorKeys.publicKey,
-        claimHandleProof,
-        claimHandlePayload
-      );
-      const calls = [createSponsoredAccountWithDelegation, claimHandle];
-
-      const payWithCapacityBatchAllOp = ExtrinsicHelper.payWithCapacityBatchAll(capacityProviderKeys, calls);
-
-      const { target: batchCompletedEvent, eventMap } = await payWithCapacityBatchAllOp.signAndSend();
-
-      if (batchCompletedEvent && !ExtrinsicHelper.api.events.utility.BatchCompleted.is(batchCompletedEvent)) {
-        assert.fail('should return a BatchCompletedEvent');
-      }
-
-      assert.notEqual(eventMap['msa.DelegationGranted'], undefined, 'should have returned DelegationGranted event');
-      assert.notEqual(eventMap['handles.HandleClaimed'], undefined, 'should have returned HandleClaimed event');
-    });
-
-    it('batch fails if one transaction fails - [createSponsoredAccountWithDelegation, claimHandle]', async function () {
-      await assert.doesNotReject(stakeToProvider(fundingSource, fundingSource, capacityProvider, amountStaked));
-
-      const addProviderPayload = await generateDelegationPayload({ ...defaultPayload });
-      const addProviderData = ExtrinsicHelper.api.registry.createType('PalletMsaAddProvider', addProviderPayload);
-      const delegatorKeys = createKeys('delegatorKeys');
-      const createSponsoredAccountWithDelegation = ExtrinsicHelper.api.tx.msa.createSponsoredAccountWithDelegation(
-        delegatorKeys.publicKey,
-        signPayloadSr25519(delegatorKeys, addProviderData),
-        addProviderPayload
-      );
-
-      const handle = 'test_handle_that_exceeds_the_byte_limit';
-      const handle_vec = new Bytes(ExtrinsicHelper.api.registry, handle);
-      const expiration = (await getBlockNumber()) + 5;
-      const handlePayload = {
-        baseHandle: handle_vec,
-        expiration: expiration,
-      };
-      const claimHandlePayload: any = ExtrinsicHelper.api.registry.createType(
-        'CommonPrimitivesHandlesClaimHandlePayload',
-        handlePayload
-      );
-      const calimHandleProof = {
-        Sr25519: u8aToHex(delegatorKeys.sign(u8aWrapBytes(claimHandlePayload.toU8a()))),
-      };
-
-      const claimHandle = ExtrinsicHelper.api.tx.handles.claimHandle(
-        delegatorKeys.publicKey,
-        calimHandleProof,
-        claimHandlePayload
-      );
-      const calls = [createSponsoredAccountWithDelegation, claimHandle];
-
-      const payWithCapacityBatchAllOp = ExtrinsicHelper.payWithCapacityBatchAll(capacityProviderKeys, calls);
-
-      await assert.rejects(payWithCapacityBatchAllOp.signAndSend(), {
-        name: 'InvalidHandleByteLength',
-      });
-    });
-  });
 });
